MdeModulePkg:Change the type of BootNext
Currently the invalid boot next set to the number of boot option,
when add a new boot option,also need update the boot next value,
otherwise it will be incorrect.Now set the type of BootNext value
to UINT32,the number out of the range of UINT16 means it is an
invalid BootNext Value.

Contributed-under: TianoCore Contribution Agreement 1.0

git-svn-id: https://svn.code.sf.net/p/edk2/code/trunk/edk2@19591 6f19259b-4bc3-4df7-8a09-765794883524

+++ MdeModulePkg/Library/BootMaintenanceManagerLib/FormGuid.h
@@ -103,7 +103,7 @@ typedef struct {
   // for Timeout, BootNext, Variables respectively
   //
   UINT16  BootTimeOut;
-  UINT16  BootNext;
+  UINT32  BootNext;
 
   //
   // This is the COM1 Attributes value storage
